fix: update model project.csm to SDK version 10.0
This makes the project.csm consistent with the rest of the model. The whole project was updated in December 2023 to SDK version 10.0, but not the project.csm file which was left in SDK version 8.1. This resulted in an inconsistent project state and a build error when executing `csm flow` using SDK version >=11.0. This is now corrected.